requirements.txt
# --- Core ---
streamlit==1.38.0
python-dotenv==1.0.1
requests==2.32.3
pandas==2.2.3
numpy==1.26.4
tqdm==4.66.5
protobuf<6,>=3.20

# --- LangChain / OpenAI / Vector store ---
langchain==0.3.7
langchain-community==0.3.7
langchain-openai==0.2.3
langchain-chroma==0.1.4
chromadb==0.5.5
faiss-cpu==1.8.0
transformers==4.44.2
sentence-transformers==3.1.0


# --- Utilities ---
python-dotenv==1.0.1
PyMuPDF==1.24.7
pypdf==4.3.1
docx2txt==0.8
tiktoken==0.7.0

# --- Optional (for debugging / UI) ---
tenacity==8.4.2
altair==5.3.0
